优化部分代码

+++ config_loader.py
@@ -8,8 +8,8 @@ from pathlib import Path
 import subprocess
 import requests
 from rich.console import Console
-from rich.panel import Panel
 from rich.progress import Progress, SpinnerColumn, TextColumn
+from rich.table import Table
 
 console = Console()
 JAR_START_TIMEOUT = 7
@@ -19,15 +19,15 @@ ARCHIVE_EXT = ('.zip', '.tar', '.tar.gz', '.tgz', '.tar.bz2', '.tbz', '.gz', '.b
 
 
 class ConfigInfo:
-    __slots__ = ('_package_file_path', '_jar_name', '_is_exists_config', '_jar_extra_args', '_health_url')
+    __slots__ = ('_package_file_path', '_jar_name', '_is_exists_config', '_jvm_args', '_health_url')
 
-    def __init__(self, package_file_path: str, jar_name: str, jar_extra_args=None, health_url=None, *,
+    def __init__(self, package_file_path: str, jar_name: str, jvm_args=None, health_url=None, *,
                  is_exists_config: bool = False):
-        if jar_extra_args is None:
-            jar_extra_args = []
+        if jvm_args is None:
+            jvm_args = []
         self._package_file_path: str = package_file_path
         self._jar_name: str = jar_name
-        self._jar_extra_args = jar_extra_args
+        self._jvm_args = jvm_args
         self._health_url = health_url
         self._is_exists_config: bool = is_exists_config
 
@@ -44,8 +44,8 @@ class ConfigInfo:
         return self._jar_name
 
     @property
-    def jar_extra_args(self):
-        return self._jar_extra_args
+    def jvm_args(self):
+        return self._jvm_args
 
     @property
     def health_url(self):
@@ -75,6 +75,11 @@ def get_full_jar_path(config_info: ConfigInfo):
 
 
 def server_is_online(health_url: str) -> bool:
+    """
+    检查服务是否在线
+    :param health_url: 健康检查URL
+    :return: True表示服务在线，False表示服务离线或无法连接
+    """
     try:
         res = requests.get(health_url, timeout=3)
         return res.status_code == 200
@@ -82,6 +87,7 @@ def server_is_online(health_url: str) -> bool:
         return False
     except requests.exceptions.RequestException:
         return False
+
 
 def get_pid_by_server_name(service_name: str) -> int:
     """
@@ -98,6 +104,10 @@ def get_pid_by_server_name(service_name: str) -> int:
 
 
 def get_config_info() -> ConfigInfo:
+    """
+    从配置文件中获取服务配置信息
+    :return: ConfigInfo对象，包含服务程序路径和名称
+    """
     if os.path.exists(CONFIG_JSON_FILE):
         with open(CONFIG_JSON_FILE, 'r', encoding="utf-8") as json_file:
             json_str = json_file.read()
@@ -117,9 +127,9 @@ def get_config_info() -> ConfigInfo:
                 if not server_jar.lower().endswith('.jar'):
                     console.print(f"服务程序名称 {server_jar} 不是支持的 jar 格式，请检查配置文件", style='bold red')
                     return ConfigInfo(server_package, '')
-                jar_extra_args = server_config.get('jar_extra_args', [])
-                health_url = server_config.get('health_url')
-                return ConfigInfo(server_package, server_jar, jar_extra_args, health_url, is_exists_config=True)
+                jvm_args = server_config.get('jvm_args', [])
+                health_url = server_config.get('health_check_url')
+                return ConfigInfo(server_package, server_jar, jvm_args, health_url, is_exists_config=True)
             else:
                 console.print("配置文件为空，请检查配置文件", style='bold red')
                 return ConfigInfo('', '')
@@ -129,6 +139,11 @@ def get_config_info() -> ConfigInfo:
 
 
 def unpack_server_package_check(config_info: ConfigInfo) -> bool:
+    """
+     解压服务程序包并校验
+    :param config_info: ConfigInfo对象，包含服务程序路径和名称
+    :return: True表示已解压，False表示未解压或解压失败
+    """
     if not os.path.exists(config_info.package_file_path):
         console.print(f"服务程序文件 {config_info.package_file_path} 不存在，请检查配置文件", style='bold red')
         return False
@@ -155,7 +170,7 @@ def unpack_server_package_check(config_info: ConfigInfo) -> bool:
     return True
 
 
-def server_is_online_by_pid_file(config_info: ConfigInfo, jar_parent_path: str, jar_file_path: str) -> bool:
+def server_is_online_by_pid_file(config_info: ConfigInfo, jar_parent_path: str) -> bool:
     # 先判断是否存在PID文件，如果有，检测服务是否存在
     path = Path(jar_parent_path)
     files = [file for file in path.iterdir() if file.is_file() and file.name.lower().endswith('.pid')]
@@ -173,14 +188,12 @@ def server_is_online_by_pid_file(config_info: ConfigInfo, jar_parent_path: str, 
                 if is_running(int(pid)):
                     console.print(f"服务已启动，pid: {pid}", style='bold yellow')
                     return True
-
-    if not os.path.isfile(jar_file_path):
-        console.print(f"启动失败 {jar_file_path} 不存在", style='bold red')
-        return False
-    return True
+    return False
 
 
 def is_running(pid: int) -> bool:
+    if pid == -1:
+        return False
     try:
         os.kill(pid, 0)
         return True
@@ -194,11 +207,35 @@ def write_pid(pid: int, jar_parent_path: str):
         pid_file.write(str(pid))
 
 
+def read_pid_file(jar_parent_path: str, jar_name: str) -> int:
+    pid = -1
+    if os.path.isfile(os.path.join(jar_parent_path, SERVER_PID_FILE)):
+        with open(os.path.join(jar_parent_path, SERVER_PID_FILE), 'r') as pid_file:
+            pid = int(pid_file.read())
+    if pid > -1:
+        return pid
+    else:
+        return get_pid_by_server_name(jar_name)
+
+
+def print_server_status(jar_name: str, pid: int, server_status: str):
+    pid = -1 if server_status == "OFFLINE" else pid
+    table = Table(show_header=True, header_style="bold magenta")
+    table.add_column("SERVER_NAME", justify="center")
+    table.add_column("PID", justify="center")
+    table.add_column("STATUS", justify="center")
+    table.add_row(jar_name.replace(".jar", "").upper(),
+                  "NONE" if pid == -1 else str(pid),
+                  server_status)
+    console.print(f'================服务状态===================')
+    console.print(table)
+
+
 def run_server(config_info: ConfigInfo, jar_parent_path: str) -> bool:
     with Progress(SpinnerColumn(), TextColumn("[process.description]{task.description}]"),
                   transient=True, console=console) as prog:
         prog.add_task(f"正在启动服务 {config_info.jar_name} ...", total=None)
-        server_process = subprocess.Popen(["java", "-jar", *config_info.jar_extra_args, config_info.jar_name],
+        server_process = subprocess.Popen(["java", *config_info.jvm_args, "-jar", config_info.jar_name],
                                           cwd=jar_parent_path,
                                           stdout=subprocess.PIPE,
                                           stderr=subprocess.STDOUT,
@@ -206,17 +243,10 @@ def run_server(config_info: ConfigInfo, jar_parent_path: str) -> bool:
         time.sleep(JAR_START_TIMEOUT)
         pid = server_process.pid
         if server_process.poll() is None:
-            console.print(Panel(
-                f"[bold green]服务已启动！[/]\n\n"
-                f"  服务名：[bold]{config_info.jar_name}[/]\n"
-                f"  PID   ：[bold]{pid}[/]\n",
-                title="[green]✔ 启动成功[/]",
-                border_style="green"
-            ))
+            print_server_status(jar_name=config_info.jar_name, pid=pid, server_status="ONLINE")
+            write_pid(pid, jar_parent_path)
+            return True
         else:
             stdout, _ = server_process.communicate()
             console.print(f"服务启动失败，错误信息：\n{stdout}", style='bold red', markup=False)
             return False
-        console.print(f"服务启动成功，进程ID：{pid}", style='bold green')
-        write_pid(pid, jar_parent_path)
-        return True
